Hide some block size details behind methods
It would be nice to implement different options for block size
information without introducing any new classes. To do that, we need to
abstract away all the information that is needed in various parts of the
codebase. This is just an early step. A lot more work is required.

Also de-duplicate SpMV code for CSBR matrices.